Fix[navbarSection, ResultsAndOrder]: refactorizo resultsandorder, actualizo dependencias, corrijo navbar, exporto interfaz
