+++ package.json
@@ -15,10 +15,11 @@
     "@reduxjs/toolkit": "^2.2.7",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "axios": "^1.7.7",
-    "next": "14.2.11",
+    "lottie-web": "^5.12.2",
+    "next": "^15.0.3",
     "notistack": "^3.0.1",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
     "react-lottie": "^1.2.4",
     "react-redux": "^9.1.2",
