+++ src/types/StoreOfProducts.ts
@@ -1,0 +1,4 @@
+export interface ResultsAndOrderProps {
+    searchType: number | null;
+    slug: string;
+}
